# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LMS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Courses
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List community posts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostWithAuthor"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostWithAuthor"
  /posts/{id}:
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Toggle like on a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"

  # Comments
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentWithAuthor"
    post:
      operationId: createComment
      tags: [comments]
      summary: Create comment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentWithAuthor"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Dashboard
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
        lastLogin:
          type: ["string", "null"]
        lastLogout:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, email, name, role, createdAt]
    SocialLinks:
      type: object
      properties:
        facebook:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        tiktok:
          type: ["string", "null"]
        twitter:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        linkedin:
          type: ["string", "null"]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
    UpdateRoleBody:
      type: object
      properties:
        role:
          type: string
      required: [role]
...
    PostWithAuthor:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        content:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        isLiked:
          type: boolean
        createdAt:
          type: string
        author:
          $ref: "#/components/schemas/User"
      required: [id, userId, content, likeCount, commentCount, isLiked, createdAt, author]
    CreatePostBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
      required: [liked, likeCount]
    CommentWithAuthor:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
        author:
          $ref: "#/components/schemas/User"
      required: [id, postId, userId, comment, createdAt, author]
    CreateCommentBody:
      type: object
      properties:
        comment:
          type: string
      required: [comment]
...