# lib/api-zod/src/generated/types/postWithAuthor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface PostWithAuthor {
  id: number;
  userId: number;
  content: string;
  likeCount: number;
  commentCount: number;
  isLiked: boolean;
  createdAt: string;
  author: User;
}