# lib/db/src/schema/users.ts
import { pgTable, text, serial, timestamp, jsonb } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const usersTable = pgTable("users", {
  id: serial("id").primaryKey(),
  email: text("email").notNull().unique(),
  passwordHash: text("password_hash").notNull(),
  name: text("name").notNull(),
  role: text("role").notNull().default("member"),
  avatar: text("avatar"),
  bio: text("bio"),
  socialLinks: jsonb("social_links").$type<{
    facebook?: string | null;
    instagram?: string | null;
    tiktok?: string | null;
    twitter?: string | null;
    whatsapp?: string | null;
    linkedin?: string | null;
  }>(),
  lastLogin: timestamp("last_login", { withTimezone: true }),
  lastLogout: timestamp("last_logout", { withTimezone: true }),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertUserSchema = createInsertSchema(usersTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertUser = z.infer<typeof insertUserSchema>;
export type User = typeof usersTable.$inferSelect;